# tests: fix freezegun's default module ignore list