--- a/tests/__init__.py
+++ b/tests/__init__.py
@@ -1,6 +1,7 @@
 import os
 import signal
 
+import freezegun.config
 import pytest
 
 # import streamlink_cli as early as possible to execute its default signal overrides
@@ -11,6 +12,12 @@ import streamlink_cli  # noqa: F401
 # immediately restore default signal handlers for the test runner
 signal.signal(signal.SIGINT, signal.default_int_handler)
 signal.signal(signal.SIGTERM, signal.default_int_handler)
+
+
+# make freezegun ignore the following pytest modules, so it doesn't mess with pytest's --duration=N report
+# when freezing time in fixtures, even though ["_pytest.runner.", "_pytest.terminal."] is already included
+# in freezegun's default module ignore list (notice the trailing dots).
+freezegun.config.configure(extend_ignore_list=["_pytest.runner", "_pytest.terminal"])
 
 
 # make pytest rewrite assertions in dynamically parametrized plugin tests