@bastimeyer 2022-12-13
- https://github.com/streamlink/streamlink/blob/c8a707e7a9f27212c0c86a262f9431c349581d19/tests/utils/test_formatter.py#L23-L37
- https://github.com/streamlink/streamlink/blob/c8a707e7a9f27212c0c86a262f9431c349581d19/tests/cli/utils/test_formatter.py#L30-L44
- https://github.com/streamlink/streamlink/blob/c8a707e7a9f27212c0c86a262f9431c349581d19/tests/plugins/test_filmon.py#L43-L49